Add Ubuntu 16.04 support, by build separate so with ffmpeg 2.8
