+++ tensorflow_io/video/BUILD
@@ -8,16 +8,37 @@ filegroup(
 )
 
 cc_binary(
-    name = 'python/ops/_video_ops.so',
+    name = 'python/ops/_video_ops_ffmpeg_3.4.so',
     srcs = [
+        "kernels/ffmpeg_3_4.cc",
+        "kernels/video_reader.h",
         "kernels/video_dataset_ops.cc",
         "ops/dataset_ops.cc",
     ],
+    includes = ["."],
     linkshared = 1,
     deps = [
         "@local_config_tf//:libtensorflow_framework",
         "@local_config_tf//:tf_header_lib",
-	"@ffmpeg//:ffmpeg",
+	"@ffmpeg_3_4//:ffmpeg",
+    ],
+    copts = ["-pthread", "-std=c++11", "-D_GLIBCXX_USE_CXX11_ABI=0", "-DNDEBUG"]
+)
+
+cc_binary(
+    name = 'python/ops/_video_ops_ffmpeg_2.8.so',
+    srcs = [
+        "kernels/ffmpeg_2_8.cc",
+        "kernels/video_reader.h",
+        "kernels/video_dataset_ops.cc",
+        "ops/dataset_ops.cc",
+    ],
+    includes = ["."],
+    linkshared = 1,
+    deps = [
+        "@local_config_tf//:libtensorflow_framework",
+        "@local_config_tf//:tf_header_lib",
+	"@ffmpeg_2_8//:ffmpeg",
     ],
     copts = ["-pthread", "-std=c++11", "-D_GLIBCXX_USE_CXX11_ABI=0", "-DNDEBUG"]
 )
@@ -28,7 +49,8 @@ py_library(
         "python/ops/video_dataset_ops.py",
     ]),
     data = [
-        ":python/ops/_video_ops.so"
+        ":python/ops/_video_ops_ffmpeg_3.4.so",
+        ":python/ops/_video_ops_ffmpeg_2.8.so",
     ],
     srcs_version = "PY2AND3",
 )
